refactor: #996 TutorialOverlay.svelte の責務分離（511行→4モジュール）
- useStepHighlight.svelte.ts: DOM要素検索・MutationObserver待機・スクロール制御 (87行)
- tutorial-step-controller.svelte.ts: ステップ追従state・$effect群・overlay操作 (155行)
- TutorialQuickCompleteDialog.svelte: resume/quickComplete/exitConfirmダイアログUI (277行)
- TutorialOverlay.svelte: オーケストレーション+スポットライト描画 (132行)
